Move ExplainPlan and ExplainPlanAttributes classes to processing module for reuse.
