+++ sql/src/main/java/org/apache/druid/sql/client/BrokerClient.java
@@ -20,7 +20,7 @@
 package org.apache.druid.sql.client;
 
 import com.google.common.util.concurrent.ListenableFuture;
-import org.apache.druid.sql.http.ExplainPlan;
+import org.apache.druid.query.explain.ExplainPlan;
 import org.apache.druid.sql.http.SqlQuery;
 import org.apache.druid.sql.http.SqlTaskStatus;
